feat: restructure ProductDescription component and add related subcomponents
