+++ src/components/sections/Catalogue/CurrentProductSection/components/Articles/ProductDescription/AddToCartButton.tsx
@@ -1,0 +1,7 @@
+export default function AddToCartButton() {
+  return (
+    <button className='rounded-lg bg-black px-6 py-3 text-white'>
+      Add to Cart
+    </button>
+  )
+}
